requirements.txt
fastapi~=0.95.0
python-dotenv~=1.0.0
orjson~=3.8.8
tortoise-orm[asyncpg]
aerich~=0.7.1
uvicorn
PyJWT~=2.6.0
jinja2~=3.1.2
itsdangerous~=2.1.2
gitpython~=3.1.31
docker~=6.0.1